database-prep/requirements.txt
pymysql==1.0.2
pandas==1.3.5
openpyxl==3.0.9
requests==2.26.0
python-dotenv==0.19.2